# docs/swagger.yaml
definitions:
...
  common.AssetType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    type: integer
    x-enum-varnames:
    - ASSET_TYPE_CRYPTO
    - ASSET_TYPE_FIAT
    - ASSET_TYPE_CARD_PRODUCT
    - ASSET_TYPE_POINT
    - ASSET_TYPE_AUTO_YIELD
...
  common.KYCLevel:
    enum:
    - "0"
    - "1"
    - "2"
    - "3"
    type: string
    x-enum-varnames:
    - KYC_LEVEL_0
    - KYC_LEVEL_1
    - KYC_LEVEL_2
    - KYC_LEVEL_3
...
  common.RatePurpose:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    - 13
    - 14
    type: integer
    x-enum-varnames:
    - RATE_PURPOSE_APPLY
    - RATE_PURPOSE_EXCHANGE
    - RATE_PURPOSE_TRANSFER
    - RATE_PURPOSE_CARD_TO_CARD
    - RATE_PURPOSE_TOP_UP
    - RATE_PURPOSE_TOP_DOWN
    - RATE_PURPOSE_FEE
    - RATE_PURPOSE_LIMIT
    - RATE_PURPOSE_MERCHANT_EXCHANGE
    - RATE_PURPOSE_MERCHANT_TRANSFER
    - RATE_PURPOSE_DISCOUNT
    - RATE_PURPOSE_DISPLAY
    - RATE_PURPOSE_CHIPPAY_EXPRESS_BUY
    - RATE_PURPOSE_TOP_UP_POINT
...
  common.UserBlockStatus:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - USER_BLOCK_STATUS_NORMAL
    - USER_BLOCK_STATUS_BLOCK
...
  shared-modules_entities.AssetVO:
    properties:
      amount:
        type: number
      categoryId:
        example: "0"
        type: string
      createdAt:
        example: "0"
        type: string
      currency:
        type: string
      currencyType:
        type: string
      frozenAmount:
        type: number
      id:
        example: "0"
        type: string
      type:
        type: string
      updatedAt:
        example: "0"
        type: string
      userId:
        example: "0"
        type: string
    type: object
...
  shared-modules_entities.DeviceInfo:
    properties:
      deviceCount:
        type: integer
      deviceID:
        type: string
      deviceName:
        type: string
      devicePasskey:
        type: boolean
      hasPasskey:
        type: boolean
      isMaster:
        type: boolean
      platform:
        type: string
    type: object
...
  shared-modules_entities.ExchangeConfirmForm:
    properties:
      exchangeKey:
        type: string
      pinCode:
        type: string
    required:
    - exchangeKey
    type: object
  shared-modules_entities.ExchangeConfirmVO:
    type: object
  shared-modules_entities.ExchangePreviewForm:
    properties:
      fromAmount:
        type: number
      fromCardId:
        example: "0"
        type: string
      toAmount:
        type: number
      toCardId:
        example: "0"
        type: string
      toCategory:
        type: string
    required:
    - fromCardId
    type: object
  shared-modules_entities.ExchangePreviewVO:
    properties:
      displayRate:
        $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
      exchangeFee:
        type: string
      expiredAt:
        example: "0"
        type: string
      fromAmount:
        type: string
      fromCardID:
        example: "0"
        type: string
      fromCategory:
        type: string
      fromCategoryID:
        example: "0"
        type: string
      fromCurrency:
        type: string
      key:
        type: string
      rate:
        items:
          $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
        type: array
      toAmount:
        type: string
      toCardID:
        example: "0"
        type: string
      toCategory:
        type: string
      toCategoryID:
        example: "0"
        type: string
      toCurrency:
        type: string
    type: object
  shared-modules_entities.ExchangeRateVO:
    properties:
      baseCurrency:
        type: string
      purpose:
        type: string
      quoteCurrency:
        type: string
      rate:
        type: number
      timestamp:
        example: "0"
        type: string
    type: object
...
  shared-modules_entities.GetEquivalentValueForm:
    properties:
      currency:
        type: string
      purpose:
        $ref: '#/definitions/common.RatePurpose'
      type:
        $ref: '#/definitions/common.AssetType'
    required:
    - currency
    type: object
  shared-modules_entities.GetEquivalentValueVO:
    properties:
      amount:
        type: number
      autualRates:
        items:
          $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
        type: array
      fees:
        additionalProperties:
          type: number
        type: object
      frozenAmount:
        type: number
      rates:
        items:
          $ref: '#/definitions/shared-modules_entities.ExchangeRateVO'
        type: array
    type: object
...
  shared-modules_entities.ListAssetsForm:
    properties:
      categoryId:
        example: "0"
        type: string
      categoryIdIn:
        items:
          type: integer
        type: array
      id:
        example: "0"
        type: string
      idIn:
        items:
          type: integer
        type: array
      typeIn:
        items:
          $ref: '#/definitions/common.AssetType'
        type: array
      userId:
        example: "0"
        type: string
    type: object
  shared-modules_entities.ListAssetsVO:
    properties:
      records:
        items:
          $ref: '#/definitions/shared-modules_entities.AssetVO'
        type: array
    type: object
...
  shared-modules_entities.LoginOrCreateForm:
    properties:
      deviceName:
        type: string
      email:
        type: string
      otpCode:
        type: string
      promotionCode:
        type: string
    required:
    - email
    - otpCode
    type: object
  shared-modules_entities.LoginOrCreateVO:
    properties:
      blockStatus:
        $ref: '#/definitions/common.UserBlockStatus'
      channel:
        type: string
      countryCode:
        description: 可能為空
        type: integer
      createdAt:
        type: integer
      device:
        $ref: '#/definitions/shared-modules_entities.DeviceInfo'
      ePoint:
        type: number
      ePointCardId:
        example: "0"
        type: string
      ePointLevel:
        type: string
      email:
        type: string
      expiredAt:
        example: "0"
        type: string
      gender:
        description: 可能為空
        type: integer
      hasPinCode:
        type: boolean
      id:
        example: "0"
        type: string
      kycLevel:
        $ref: '#/definitions/common.KYCLevel'
      phoneNumber:
        description: 可能為空
        type: string
      token:
        type: string
      updatedAt:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /web/account/assets/getEquivalent:
    post:
      description: list assets
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.GetEquivalentValueForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.GetEquivalentValueVO'
      tags:
      - web/account
  /web/account/assets/list:
    post:
      description: list assets
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.ListAssetsForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ListAssetsVO'
      tags:
      - web/account
...
  /web/exchange/confirm:
    post:
      description: Apply for a new exchange confirmation.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.ExchangeConfirmForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ExchangeConfirmVO'
      summary: Apply for a new exchange confirmation.
      tags:
      - web/exchange
  /web/exchange/preview:
    post:
      description: Apply for a new ex preview.
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.ExchangePreviewForm'
      - description: User token
        in: header
        name: X-Token
        required: true
        type: string
      - description: accept language
        in: header
        name: Accept-Language
        type: string
      - description: Extend
        in: header
        name: X-Extend
        type: string
      - description: Convert
        in: header
        name: X-Convert
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.ExchangePreviewVO'
      summary: apply for a new ex preview. 兌換(deposit充值) 自己的數幣轉數幣
      tags:
      - web/exchange
...
  /web/user/loginOrRegister:
    post:
      description: Login or register a user. 如果有pin code，hasPinCode會是true
      parameters:
      - description: body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/shared-modules_entities.LoginOrCreateForm'
      - description: 回傳User token
        in: header
        name: X-Token
        type: string
      - description: 回傳失效時間
        in: header
        name: X-ExpiredTs
        type: string
      responses:
        "0":
          description: data
          schema:
            $ref: '#/definitions/shared-modules_entities.LoginOrCreateVO'
      tags:
      - web/user
...
  /web/user/logout:
    post:
      consumes:
      - application/json
      description: Logout the user by invalidating their token
      parameters:
      - description: User Token
        in: header
        name: X-Token
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Logout
      tags:
      - web/user
...
swagger: "2.0"